+++ package.json
@@ -13,6 +13,7 @@
     "@baronha/react-native-multiple-image-picker": "^0.4.13",
     "@react-native-async-storage/async-storage": "^1.17.11",
     "@react-native-community/google-signin": "^5.0.0",
+    "@react-native-community/viewpager": "^5.0.11",
     "@react-native-seoul/kakao-login": "^5.2.6",
     "@react-navigation/bottom-tabs": "^6.5.4",
     "@react-navigation/material-top-tabs": "^6.6.0",
@@ -25,13 +26,14 @@
     "react-native": "0.69.8",
     "react-native-dotenv": "^3.4.7",
     "react-native-fast-image": "^8.6.3",
+    "react-native-pager-view": "^6.1.4",
     "react-native-paper": "^5.2.0",
     "react-native-permissions": "^3.6.1",
     "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "^3.19.0",
     "react-native-splash-screen": "^3.3.0",
     "react-native-svg": "^13.8.0",
-    "react-native-tab-view": "^3.4.0",
+    "react-native-tab-view": "^3.5.0",
     "react-redux": "^8.0.5"
   },
   "devDependencies": {
